+++ src/components/Politician.vue
@@ -33,43 +33,14 @@
         <span class="block text-sm capitalize">{{position}}</span>
       </div>
     </div>
-
-    <div class="absolute w-full h-full top-0 left-0">
-      <div class="inner w-full h-full text-right">
-        <our-dropdown class="inline-block mt-5 bg-white font-circular" list-margin="-mt-6" padding="py-1" width="w-64" headingFont="font-circular" leaveOpen listClass="leaders-dropdown-list">
-          <template v-slot:heading>
-            <button class="btn-subscribe px-4 py-1" :class="{ 'active': subscribed }">
-              <span class="align-middle" v-if="!subscribed">Subscribe</span>
-              <span class="align-middle" v-if="subscribed">Subscribed</span>
-              <span class="align-middle loading inline sm ml-2" v-if="processing"></span>
-              <img class="ml-2 align-baseline" v-else-if="subscribed" src="@/assets/img/green-tick.svg"/>
-            </button>
-          </template>
-          <our-dropdown-item>
-            <label class="flex justify-between text-xs items-center">
-              Be notified on this platform
-              <our-checkbox :model="subscription.feeds" :disabled="processing" v-on:change="changeSubscription('feeds', $event)"></our-checkbox>
-            </label>
-          </our-dropdown-item>
-          <our-dropdown-item>
-            <label class="flex justify-between text-xs items-center">
-              Receive email notifications
-              <our-checkbox :model="subscription.email" :disabled="processing" v-on:change="changeSubscription('email', $event)"></our-checkbox>
-            </label>
-          </our-dropdown-item>
-          <our-dropdown-divider v-show="subscribed"/>
-          <our-dropdown-item>
-            <a v-show="subscribed" @click="removeSubscription()" class="text-xs text-gray-400 w-full" :class="unsubscribeLinkClass">Unsubscribe from this leader</a>
-          </our-dropdown-item>
-        </our-dropdown>
-      </div>
+    <div class="absolute w-full h-full top-0 left-0 hidden-overflow-transparent">
+      <our-politician-subscription :politicianSubscriptions="subscriptions" :politician="politician"></our-politician-subscription>
     </div>
   </div>
 </div>
 </template>
 
 <script>
-import { mapActions } from 'vuex';
 import StringUtil from '@/helpers/stringUtil';
 
 export default {
@@ -84,6 +55,7 @@ export default {
     },
     subscriptions: {
       type: Array,
+      default: () => ([]),
     },
   },
   computed: {
@@ -94,76 +66,6 @@ export default {
         return StringUtil.getPoliticalPosition(background);
       }
       return 'N/A';
-    },
-    unsubscribeLinkClass() {
-      if (this.processing) {
-        return 'cursor-not-allowed text-gray-300';
-      }
-      return '';
-    },
-    subscribed() {
-      return this.subscription.feeds || this.subscription.email;
-    },
-  },
-  data() {
-    return {
-      processing: false,
-      subscription: {
-        feeds: !!this.getSubscriptionIdByType('feeds'),
-        email: !!this.getSubscriptionIdByType('email'),
-      },
-      subscriptionsServices: this.$serviceFactory.subscriptions,
-    };
-  },
-  methods: {
-    ...mapActions([
-      'displayError',
-    ]),
-    async addSubscription(type) {
-      try {
-        this.processing = true;
-        await this.subscriptionsServices.addSubscription({
-          politicianId: this.politician.id,
-          type,
-        });
-        this.subscription[type] = true;
-        this.processing = false;
-      } catch (error) {
-        this.subscription[type] = false;
-        this.processing = false;
-        this.displayError(error);
-      }
-    },
-    async removeSubscription(type) {
-      try {
-        this.processing = true;
-        this.subscription[type] = true;
-        if (type) {
-          await this.subscriptionsServices.removeSubscription(this.getSubscriptionIdByType(type));
-          this.subscription[type] = false;
-        } else {
-          await this.subscriptionsServices.removePoliticianSubscriptions(this.politician.id);
-          this.subscription.feeds = false;
-          this.subscription.email = false;
-        }
-
-        this.processing = false;
-      } catch (error) {
-        this.subscription[type] = true;
-        this.processing = false;
-        this.displayError(error);
-      }
-    },
-    async changeSubscription(type, isSubscribed) {
-      if (isSubscribed) {
-        this.addSubscription(type);
-      } else {
-        this.removeSubscription(type);
-      }
-    },
-    getSubscriptionIdByType(type) {
-      const subscription = this.subscriptions.find(sub => sub.type === type);
-      return subscription ? subscription.id : null;
     },
   },
 };
