+++ src/components/index.js
@@ -17,6 +17,7 @@ import Offices from '@/components/Offices.vue';
 import Pagination from '@/components/Pagination.vue';
 import PoliticalParty from '@/components/PoliticalParty.vue';
 import Politician from '@/components/Politician.vue';
+import PoliticianSubscription from '@/components/PoliticianSubscription.vue';
 import QuarterlyView from '@/components/QuarterlyView.vue';
 import SideScroll from '@/components/SideScroll.vue';
 import SocialLogin from '@/components/SocialLogin.vue';
@@ -37,6 +38,7 @@ Vue.component('our-offices', Offices);
 Vue.component('our-pagination', Pagination);
 Vue.component('our-political-party', PoliticalParty);
 Vue.component('our-politician', Politician);
+Vue.component('our-politician-subscription', PoliticianSubscription);
 Vue.component('our-quarterly-view', QuarterlyView);
 Vue.component('our-select-dropdown', SelectDropdown);
 Vue.component('our-social-login', SocialLogin);
